feat: 処理の切り出し
