+++ src/pages/calc-stats/index.tsx
@@ -1,24 +1,23 @@
 import { NextPage } from 'next'
 import { useRecoilState } from 'recoil'
 import { CalcStatsTemplate } from '@/components/templates/CalcStatsTemplate'
-import { levelState, selectedNatureState, selectedPokemonState, statsState } from '@/store'
-import { Nature, PokemonData, Stat } from '@/types'
+import { levelState, selectedNatureState, selectedPokemonState } from '@/store'
+import { Nature, PokemonData } from '@/types'
+import { useStatsMutators, useStatsState } from '@/store/statsState'
 
 const CalcStats: NextPage = () => {
   const [selectedPokemon, setSelectedPokemon] = useRecoilState(selectedPokemonState)
   const [selectedNature, setSelectedNature] = useRecoilState(selectedNatureState)
   const [level, setLevel] = useRecoilState(levelState)
-  const [stats, setStats] = useRecoilState(statsState)
+
+  const stats = useStatsState()
+  const { updateStats } = useStatsMutators()
 
   const updatePokemon = (pokemon: PokemonData) => {
     setSelectedPokemon(pokemon)
   }
   const updateNature = (nature: Nature) => {
     setSelectedNature(nature)
-  }
-
-  const updateStats = (stats: Stat[]) => {
-    setStats(stats)
   }
 
   const updateLevel = (level: number | '') => {
