+++ src/store/index.ts
@@ -48,45 +48,3 @@ export const levelState = atom<number | ''>({
   key: 'levelState',
   default: 50,
 })
-
-export const statsState = atom<Stat[]>({
-  key: 'statsState',
-  default: [
-    {
-      name: 'ＨＰ',
-      initial: 'H',
-      individualValue: 31,
-      effortValue: '',
-    },
-    {
-      name: 'こうげき',
-      initial: 'A',
-      individualValue: 31,
-      effortValue: '',
-    },
-    {
-      name: 'ぼうぎょ',
-      initial: 'B',
-      individualValue: 31,
-      effortValue: '',
-    },
-    {
-      name: 'とくこう',
-      initial: 'C',
-      individualValue: 31,
-      effortValue: '',
-    },
-    {
-      name: 'とくぼう',
-      initial: 'D',
-      individualValue: 31,
-      effortValue: '',
-    },
-    {
-      name: 'すばやさ',
-      initial: 'S',
-      individualValue: 31,
-      effortValue: '',
-    },
-  ],
-})
